feat: factor out route
